net-switch-automation/requirements.txt
pytest
pytest-cov
pydantic>=2.0.0
jinja2
pyyaml
